+++ src/sql/run/resultset.py
@@ -1,4 +1,3 @@
-import warnings
 import re
 import operator
 from functools import reduce
@@ -14,7 +13,6 @@ from sql.column_guesser import ColumnGuesserMixin
 from sql.run.csv import CSVWriter, CSVResultDescriptor
 from sql.telemetry import telemetry
 from sql.run.table import CustomPrettyTable
-from sql.warnings import JupySQLDataFramePerformanceWarning
 
 
 class ResultSetsManager:
@@ -463,11 +461,22 @@ def unduplicate_field_names(field_names):
 def _convert_to_data_frame(
     result_set, converter_name, constructor, constructor_kwargs=None
 ):
+    """
+    Convert the result set to a pandas DataFrame, using native DuckDB methods if
+    possible
+    """
     constructor_kwargs = constructor_kwargs or {}
-    has_converter_method = hasattr(result_set.sqlaproxy, converter_name)
+
+    # maybe create accessors in the connection objects?
+    if result_set._conn.is_dbapi_connection:
+        native_connection = result_set.sqlaproxy
+    else:
+        native_connection = result_set._conn._connection.connection
+
+    has_converter_method = hasattr(native_connection, converter_name)
 
     # native duckdb connection
-    if hasattr(result_set.sqlaproxy, converter_name):
+    if has_converter_method:
         # we need to re-execute the statement because if we fetched some rows
         # already, .df() will return None. But only if it's a select statement
         # otherwise we might end up re-execute INSERT INTO or CREATE TABLE
@@ -475,36 +484,14 @@ def _convert_to_data_frame(
         is_select = _statement_is_select(result_set._statement)
 
         if is_select:
-            result_set.sqlaproxy.execute(result_set._statement)
+            native_connection.execute(result_set._statement)
 
-        return getattr(result_set.sqlaproxy, converter_name)()
+        return getattr(native_connection, converter_name)()
     else:
         frame = constructor(
             (tuple(row) for row in result_set),
             **constructor_kwargs,
         )
-
-        # NOTE: in JupySQL 0.7.9, we were opening a raw new connection so people
-        # using SQLALchemy still had the native performance to convert to data frames
-        # but this led to other problems because the native connection didn't
-        # have the same state as the SQLAlchemy connection, yielding confusing
-        # errors. So we decided to remove this and just warn the user that
-        # performance might be slow and they could use a native connection
-        if (
-            result_set._dialect == "duckdb"
-            and not has_converter_method
-            and len(frame) >= 1_000
-        ):
-            DOCS = "https://jupysql.ploomber.io/en/latest/integrations/duckdb.html"
-            WARNINGS = "https://jupysql.ploomber.io/en/latest/tutorials/duckdb-native-sqlalchemy.html#supress-warnings"  # noqa: E501
-
-            warnings.warn(
-                "It looks like you're using DuckDB with SQLAlchemy. "
-                "For faster conversions, use "
-                f" a DuckDB native connection. Docs: {DOCS}."
-                f" to suppress this warning, see: {WARNINGS}",
-                category=JupySQLDataFramePerformanceWarning,
-            )
 
         return frame