+++ src/sql/warnings.py
@@ -1,6 +1,2 @@
-class JupySQLDataFramePerformanceWarning(UserWarning):
-    pass
-
-
 class JupySQLQuotedNamedParametersWarning(UserWarning):
     pass