+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * [Fix] Fixed vertical color breaks in histograms (#702)
 * [Fix] Showing feedback when switching connections (#727)
 * [Fix] Fix error that caused some connections not to be closed when calling `--close/-x`
+* [Fix] Fix bug that caused the query transpilation process to fail when passing multiple statements
+* [Feature] Adds `with_` to `{SQLAlchemyConnection, DBAPIConnection}.raw_execute` to resolve CTEs
+* [Fix] Fixes error when creating tables and querying them in the same cell when using DuckDB + SQLAlchemy (#674)
+* [Fix] Using native methods to convert to data frames from DuckDB when using native connections and SQLAlchemy
 * [Fix] Fix error that caused literals like `':something'` to be interpreted as query parameters
 * [Feature] allows parametrizing queries with `:variable` with `%config SqlMagic.named_paramstyle = True`